Introduce execution.hpp header Reorganize execution_categories.hpp underneath execution/
