+++ agency/detail/make_tuple_if_not_scoped.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <agency/execution_categories.hpp>
+#include <agency/execution/execution_categories.hpp>
 #include <agency/detail/tuple.hpp>
 #include <utility>
