+++ agency/execution.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <agency/detail/config.hpp>
+#include <agency/execution/execution_categories.hpp>
+
