+++ agency/executor/concurrent_executor.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <agency/future.hpp>
-#include <agency/execution_categories.hpp>
+#include <agency/execution/execution_categories.hpp>
 #include <agency/detail/invoke.hpp>
 #include <agency/detail/type_traits.hpp>
